fix: fe requested fixes
