+++ components/partials/common/trade/percent-amount-options.vue
@@ -291,7 +291,7 @@ export default Vue.extend({
       isSpot
         ? this.updateSpotQuoteAmount()
         : this.$emit(
-            'update-quote-amount-from-percentage',
+            'update:quoteAmountFromPercentage',
             quoteAmountFromPercentage
           )
     },
