+++ components/partials/derivatives/trading/order-details.vue
@@ -175,7 +175,10 @@ import {
   UiDerivativeMarketWithToken,
   ZERO_IN_BASE
 } from '@injectivelabs/sdk-ui-ts'
-import { DerivativeOrderSide } from '@injectivelabs/sdk-ts'
+import {
+  DerivativeOrderSide,
+  cosmosSdkDecToBigNumber
+} from '@injectivelabs/sdk-ts'
 import Drawer from '~/components/elements/drawer.vue'
 import {
   UI_DEFAULT_AMOUNT_DISPLAY_DECIMALS,
@@ -183,6 +186,7 @@ import {
 } from '~/app/utils/constants'
 import { getDecimalsFromNumber } from '~/app/utils/helpers'
 import { Icon } from '~/types'
+import { TradingRewardsCampaign } from '~/app/client/types/exchange'
 
 export default Vue.extend({
   components: {
@@ -226,16 +230,6 @@ export default Vue.extend({
     },
 
     makerFeeRate: {
-      required: true,
-      type: Object as PropType<BigNumberInBase>
-    },
-
-    takerExpectedPts: {
-      required: true,
-      type: Object as PropType<BigNumberInBase>
-    },
-
-    makerExpectedPts: {
       required: true,
       type: Object as PropType<BigNumberInBase>
     },
@@ -302,6 +296,10 @@ export default Vue.extend({
       return this.$accessor.derivatives.market
     },
 
+    tradingRewardsCampaign(): TradingRewardsCampaign | undefined {
+      return this.$accessor.exchange.tradingRewardsCampaign
+    },
+
     totalWithFeesToFormat(): string {
       const { totalWithFees, market } = this
 
@@ -320,6 +318,131 @@ export default Vue.extend({
       }
 
       return price.toFormat(market.priceDecimals)
+    },
+
+    makerExpectedPts(): BigNumberInBase {
+      const { market, makerFeeRate, tradingRewardsCampaign, fees } = this
+
+      if (!market) {
+        return ZERO_IN_BASE
+      }
+
+      if (makerFeeRate.lte(0)) {
+        return ZERO_IN_BASE
+      }
+
+      if (
+        !tradingRewardsCampaign ||
+        !tradingRewardsCampaign.tradingRewardCampaignInfo ||
+        !tradingRewardsCampaign.tradingRewardCampaignInfo
+          .disqualifiedMarketIdsList
+      ) {
+        return ZERO_IN_BASE
+      }
+
+      const disqualified =
+        tradingRewardsCampaign.tradingRewardCampaignInfo.disqualifiedMarketIdsList.find(
+          (marketId) => marketId === market.marketId
+        )
+
+      if (disqualified) {
+        return ZERO_IN_BASE
+      }
+
+      const denomIncluded =
+        tradingRewardsCampaign.tradingRewardCampaignInfo.quoteDenomsList.find(
+          (denom) => denom === market.quoteDenom
+        )
+
+      if (!denomIncluded) {
+        return ZERO_IN_BASE
+      }
+
+      const boostedList = tradingRewardsCampaign.tradingRewardCampaignInfo
+        .tradingRewardBoostInfo
+        ? tradingRewardsCampaign.tradingRewardCampaignInfo
+            .tradingRewardBoostInfo.boostedDerivativeMarketIdsList
+        : []
+      const multipliersList = tradingRewardsCampaign.tradingRewardCampaignInfo
+        .tradingRewardBoostInfo
+        ? tradingRewardsCampaign.tradingRewardCampaignInfo
+            .tradingRewardBoostInfo.derivativeMarketMultipliersList
+        : []
+
+      const boosted = boostedList.findIndex(
+        (derivativeMarketId) => derivativeMarketId === market.marketId
+      )
+
+      const boostedMultiplier =
+        boosted >= 0
+          ? cosmosSdkDecToBigNumber(
+              multipliersList[boosted]
+                ? multipliersList[boosted].makerPointsMultiplier
+                : 1
+            )
+          : 1
+
+      return new BigNumberInBase(fees).times(boostedMultiplier)
+    },
+
+    takerExpectedPts(): BigNumberInBase {
+      const { market, tradingRewardsCampaign, fees } = this
+
+      if (!market) {
+        return ZERO_IN_BASE
+      }
+
+      if (
+        !tradingRewardsCampaign ||
+        !tradingRewardsCampaign.tradingRewardCampaignInfo ||
+        !tradingRewardsCampaign.tradingRewardCampaignInfo
+          .disqualifiedMarketIdsList
+      ) {
+        return ZERO_IN_BASE
+      }
+
+      const disqualified =
+        tradingRewardsCampaign.tradingRewardCampaignInfo.disqualifiedMarketIdsList.find(
+          (marketId) => marketId === market.marketId
+        )
+
+      if (disqualified) {
+        return ZERO_IN_BASE
+      }
+
+      const denomIncluded =
+        tradingRewardsCampaign.tradingRewardCampaignInfo.quoteDenomsList.find(
+          (denom) => denom === market.quoteDenom
+        )
+
+      if (!denomIncluded) {
+        return ZERO_IN_BASE
+      }
+
+      const boostedList = tradingRewardsCampaign.tradingRewardCampaignInfo
+        .tradingRewardBoostInfo
+        ? tradingRewardsCampaign.tradingRewardCampaignInfo
+            .tradingRewardBoostInfo.boostedDerivativeMarketIdsList
+        : []
+      const multipliersList = tradingRewardsCampaign.tradingRewardCampaignInfo
+        .tradingRewardBoostInfo
+        ? tradingRewardsCampaign.tradingRewardCampaignInfo
+            .tradingRewardBoostInfo.derivativeMarketMultipliersList
+        : []
+
+      const boosted = boostedList.findIndex(
+        (derivativeMarketId) => derivativeMarketId === market.marketId
+      )
+      const boostedMultiplier =
+        boosted >= 0
+          ? cosmosSdkDecToBigNumber(
+              multipliersList[boosted]
+                ? multipliersList[boosted].takerPointsMultiplier
+                : 1
+            )
+          : 1
+
+      return new BigNumberInBase(fees).times(boostedMultiplier)
     },
 
     notionalValueToFormat(): string {
